feat: add link role and access guard; implement CRUD operations for link management and QR code generation

+++ app/ressources/link_ressource.py
@@ -1,8 +1,12 @@
 from typing import Annotated, Literal
 from fastapi import Depends, Query, Request, Response
+from app.classes.auth_permission import Role
 from app.container import InjectInMethod
+from app.decorators.guards import AccessLinkGuard
 from app.decorators.handlers import TortoiseHandler
-from app.definition._ressource import BaseHTTPRessource, HTTPMethod, HTTPRessource, HTTPStatusCode, UseHandler, UseLimiter
+from app.decorators.permissions import JWTRouteHTTPPermission
+from app.definition._ressource import BaseHTTPRessource, HTTPMethod, HTTPRessource, HTTPStatusCode, UseGuard, UseHandler, UseLimiter, UsePermission, UseRoles
+from app.depends.dependencies import get_auth_permission
 from app.depends.my_depends import get_link
 from app.services.config_service import ConfigService
 from app.services.database_service import RedisService
@@ -13,7 +17,9 @@ from app.models.link_model import LinkORM,LinkModel, UpdateLinkModel
 
 
 LINK_MANAGER_PREFIX = 'manage'
- 
+
+@UseRoles([Role.LINK])
+@UsePermission(JWTRouteHTTPPermission)
 @UseHandler(TortoiseHandler)
 @HTTPRessource(LINK_MANAGER_PREFIX)
 class CRUDLinkRessource(BaseHTTPRessource):
@@ -25,36 +31,54 @@ class CRUDLinkRessource(BaseHTTPRessource):
         self.redisService = redisService
         self.linkService = linkService
 
+    @UseRoles([Role.ADMIN])
     @HTTPStatusCode(201)
     @BaseHTTPRessource.HTTPRoute('/', methods=[HTTPMethod.POST])
-    async def add_link(self, request: Request, linkModel: LinkModel, response: Response):
+    async def add_link(self, request: Request, linkModel: LinkModel, response: Response,authPermission=Depends(get_auth_permission)):
         link = linkModel.model_dump()
         link = await LinkORM.create(**link)
         return {"data": link.to_json, "message": "Link created successfully"}
 
+    @UseRoles([Role.PUBLIC])
     @BaseHTTPRessource.HTTPRoute('/', methods=[HTTPMethod.GET])
-    def read_link(self, request: Request, link: Annotated[LinkORM, Depends(get_link)]):
+    def read_link(self, request: Request, link: Annotated[LinkORM, Depends(get_link)],authPermission=Depends(get_auth_permission)):
         return {"data": link.to_json, "message": "Link retrieved successfully"}
 
+    @UseRoles([Role.ADMIN])
     @BaseHTTPRessource.HTTPRoute('/', methods=[HTTPMethod.DELETE])
-    async def delete_link(self, link: Annotated[LinkORM, Depends(get_link)], archive: bool = Query(False)):
+    async def delete_link(self, link: Annotated[LinkORM, Depends(get_link)], archive: bool = Query(False),authPermission=Depends(get_auth_permission)):
         link_data = link.to_json.copy()
-        await link.delete()
-        return {"data": link_data, "message": "Link deleted successfully"}
+        if not archive:
+            await link.delete()
+            return {"data": link_data, "message": "Link deleted successfully"}
+        
+        link.archived = True
+        await link.save()
+        return {"data": link_data, "message": "Link archived successfully"}
 
+    @UseRoles([Role.ADMIN])
     @HTTPStatusCode(200)
     @BaseHTTPRessource.HTTPRoute('/', methods=[HTTPMethod.PUT])
-    async def update_link(self, link: Annotated[LinkORM, Depends(get_link)], linkUpdateModel: UpdateLinkModel,response:Response):
-        link.expiration = linkUpdateModel.expiration
-        link.link_name = linkUpdateModel.link_name
+    async def update_link(self, link: Annotated[LinkORM, Depends(get_link)], linkUpdateModel: UpdateLinkModel,response:Response,authPermission=Depends(get_auth_permission)):
+        if linkUpdateModel.archived != None:
+            link.archived = linkUpdateModel.archived
+        
+        if linkUpdateModel.expiration:
+            link.expiration = linkUpdateModel.expiration
+        
+        if linkUpdateModel.link_name:
+            link.link_name = linkUpdateModel.link_name
+
         await link.save()
         return {"data": link.to_json(), "message": "Link updated successfully"}
 
-    
-    @BaseHTTPRessource.HTTPRoute('/code/',methods=[HTTPMethod.GET])
-    def get_qrcode(self,link:Annotated[LinkORM,Depends(get_link)],):
+    @UseGuard(AccessLinkGuard)
+    @UseRoles([Role.PUBLIC])
+    @BaseHTTPRessource.HTTPRoute('/code/{link_id}/',methods=[HTTPMethod.GET])
+    def get_qrcode(self,link_id:str,link:Annotated[LinkORM,Depends(get_link)],):
         ...
     
+    @UseRoles([Role.ADMIN])
     @BaseHTTPRessource.HTTPRoute('/verify',methods=[HTTPMethod.PATCH])
     def verify(self,request:Request,link:Annotated[LinkORM,Depends(get_link)],verify_type:Literal['well-known','domain']=Depends(verify_url)):
         ...
@@ -73,14 +97,14 @@ class LinkRessource(BaseHTTPRessource):
     
 
     @UseLimiter(limit_value='10000/min')
-    @BaseHTTPRessource.HTTPRoute('/visits/{url}',methods=[HTTPMethod.GET,HTTPMethod.POST],mount=True)
-    def visit_url(self,request:Request,url:str):
+    @BaseHTTPRessource.HTTPRoute('/{link_id}',methods=[HTTPMethod.GET,HTTPMethod.POST],mount=True)
+    def visit_url(self,request:Request,link:Annotated[LinkORM,Depends(get_link)]):
         
         ... 
 
     @UseLimiter(limit_value='10000/min')
-    @BaseHTTPRessource.HTTPRoute('/email-track/{url}',methods=[HTTPMethod.GET,HTTPMethod.POST],mount=True)
-    def track_email(self,request:Request,url):
+    @BaseHTTPRessource.HTTPRoute('/email-track/',methods=[HTTPMethod.GET,HTTPMethod.POST],mount=True)
+    def track_email(self,request:Request):
         ...
